Split widget: settings panel
